+++ docs/changelog.md
@@ -13,6 +13,7 @@ category: Administration
 
 * chore: include jenkins repository for external libraries
 * chore: improve Github action for continuous integration
+* test: add test creating high number of subtasks
 
 ### [25.10.0] -  2025-10-03